impl ExclusiveSystemParam for SystemName
